+++ pyanalyze/implementation.py
@@ -1,9 +1,23 @@
+import ast
 import collections
 import collections.abc
 import inspect
 import typing
 from itertools import product
-from typing import Callable, cast, Dict, NewType, Optional, Sequence, TypeVar, Union
+from typing import (
+    Callable,
+    Type,
+    cast,
+    Dict,
+    NewType,
+    Optional,
+    Sequence,
+    TypeVar,
+    Union,
+)
+
+from .annotated_types import MaxLen
+from .annotated_types import MinLen
 
 import qcore
 import typing_extensions
@@ -37,6 +51,8 @@ from .value import (
     AnnotatedValue,
     AnySource,
     AnyValue,
+    CustomCheckExtension,
+    annotate_value,
     assert_is_value,
     CallableValue,
     CanAssignContext,
@@ -1332,12 +1348,37 @@ def len_of_value(val: Value) -> Value:
     return TypedValue(int)
 
 
+def len_transformer(val: Value, op: Type[ast.AST], comparator: object) -> Value:
+    if not isinstance(comparator, int):
+        return val
+    if isinstance(len_of_value(val), KnownValue):
+        return val  # no need to specify
+    if op is ast.Eq:
+        return annotate_value(
+            val,
+            [
+                CustomCheckExtension(MinLen(comparator)),
+                CustomCheckExtension(MaxLen(comparator)),
+            ],
+        )
+    elif op is ast.Lt:
+        return annotate_value(val, [CustomCheckExtension(MaxLen(comparator - 1))])
+    elif op is ast.LtE:
+        return annotate_value(val, [CustomCheckExtension(MaxLen(comparator))])
+    elif op is ast.Gt:
+        return annotate_value(val, [CustomCheckExtension(MinLen(comparator + 1))])
+    elif op is ast.GtE:
+        return annotate_value(val, [CustomCheckExtension(MinLen(comparator))])
+    else:
+        return val
+
+
 def _len_impl(ctx: CallContext) -> ImplReturn:
     varname = ctx.varname_for_arg("obj")
     if varname is None:
         constraint = NULL_CONSTRAINT
     else:
-        constraint = PredicateProvider(varname, len_of_value)
+        constraint = PredicateProvider(varname, len_of_value, len_transformer)
     return ImplReturn(len_of_value(ctx.vars["obj"]), constraint)