+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Support inferring `MinLen` and `MaxLen` annotations based
+  on `len()` checks (#680)
 - Expose a convenience API for runtime type checking in the
   `pyanalyze.runtime` module (#674)
 - Support for annotations from the `annotated-types` library (#673)